[ADD] admin point controller
