--- src/main/java/com/server/Dotori/model/member/controller/point/PointController.java
@@ -1,30 +1,0 @@
-package com.server.Dotori.model.member.controller.point;
-
-import com.server.Dotori.model.member.service.point.PointService;
-import com.server.Dotori.response.ResponseService;
-import com.server.Dotori.response.result.CommonResult;
-import com.server.Dotori.response.result.SingleResult;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/v1")
-public class PointController {
-
-    private final PointService pointService;
-    private final ResponseService responseService;
-
-    @GetMapping("/point/{classId}")
-    @ResponseStatus( HttpStatus.OK )
-    @ApiImplicitParams({
-            @ApiImplicitParam(name = "Authorization", value = "로그인 성공 후 access_token", required = true, dataType = "String", paramType = "header"),
-            @ApiImplicitParam(name = "RefreshToken", value = "로그인 성공 후 refresh_token", required = false, dataType = "String", paramType = "header")
-    })
-    public SingleResult getAllStudentsPoint(@PathVariable("classId") Long id) {
-        return responseService.getSingleResult(pointService.getAllStudentPoint(id));
-    }
-}
