+++ src/main/java/com/server/Dotori/model/member/controller/point/developer/DeveloperPointController.java
@@ -1,0 +1,2 @@
+package com.server.Dotori.model.member.controller.point.developer;public class DeveloperPointController {
+}
